only extract exceptions when exit code is not 0

+++ worker/status_worker.go
@@ -172,7 +172,7 @@ func (sw *statusWorker) processEKSRun(run state.Run) {
 		if run.Status != updatedRun.Status && (updatedRun.PodName == run.PodName) {
 			_ = sw.log.Log("message", "updating eks run status", "pod", updatedRun.PodName, "status", updatedRun.Status, "exit_code", updatedRun.ExitCode)
 
-			if sw.exceptionExtractorClient != nil && updatedRun.ExitCode != nil && *(updatedRun.ExitCode) != 0 {
+			if sw.exceptionExtractorClient != nil && updatedRun.ExitCode != nil && *(updatedRun.ExitCode) != int64(0) {
 				jobUrl := fmt.Sprintf("%s/extract/%s", sw.exceptionExtractorUrl, run.RunID)
 				res, err := sw.exceptionExtractorClient.Get(jobUrl)
 				if err == nil {
